SkinModel, RegionAttachmentModel updated, basic tests
