+++ Package.swift
@@ -17,7 +17,8 @@ let package = Package(
                         .process("Resources/spineboy-ess.json"),
                         .process("Resources/skeleton.json"),
                         .process("Resources/bones.json"),
-                        .process("Resources/slots.json")
+                        .process("Resources/slots.json"),
+                        .process("Resources/skins.json")
                     ])
     ]
 )
