+++ Sources/Spine/Model/Skin/Attachments/AttachmentModel.swift
@@ -1,0 +1,13 @@
+//
+//  AttachmentModel.swift
+//  
+//
+//  Created by Max Gribov on 06.11.2022.
+//
+
+import Foundation
+
+protocol AttachmentModel {
+
+    var name: String { get }
+}
