+++ Sources/Spine/Model/SkinModels.swift
@@ -8,259 +8,7 @@
 
 import SpriteKit
 
-//MARK: - Skin Model
 
-struct SkinModel {
-    
-    let name: String
-    let slots: [SkinSlotModel]?
-}
-
-extension SkinModel: Decodable {
-    
-    enum Keys: String, CodingKey {
-        
-        case name
-        case attachments
-    }
-    
-    init(from decoder: Decoder) throws {
-        
-        let container = try decoder.container(keyedBy: Keys.self)
-        name = try container.decode(String.self, forKey: .name)
-        
-        let slotsContainer = try container.nestedContainer(keyedBy: SkinModel.KeysType.self, forKey: .attachments)
-        
-        var _slots = [SkinSlotModel]()
-        
-        for slotKey in slotsContainer.allKeys {
-            
-            let slotContainer = try slotsContainer.nestedContainer(keyedBy: SkinSlotModel.KeysType.self, forKey: slotKey)
-            let slot = try SkinSlotModel(slotKey.stringValue, slotContainer)
-            _slots.append(slot)
-        }
-
-        slots = _slots
-    }
-}
-
-extension SkinModel: SpineDecodableDictionary {
-    
-    typealias KeysType = SpineNameKey
-    
-    init(_ name: String, _ container: KeyedDecodingContainer<KeysType>) throws {
-        
-        var slots = [SkinSlotModel]()
-        
-        for slotKey in container.allKeys {
-            
-            let slotContainer = try container.nestedContainer(keyedBy: SkinSlotModel.KeysType.self, forKey: slotKey)
-            let slot = try SkinSlotModel(slotKey.stringValue, slotContainer)
-            slots.append(slot)
-        }
-        
-        self.name = name
-        self.slots = slots
-    }
-}
-
-//MARK: - Skin Slot Model
-
-struct SkinSlotModel {
-    
-    let name: String
-    let attachments: [AttachmentModelType]?
-}
-
-extension SkinSlotModel: SpineDecodableDictionary {
-    
-    enum SkinSlotModelDecodingError: Error {
-        
-        case attachmentTypeUnknown
-    }
-    
-    typealias KeysType = SpineNameKey
-    
-    init(_ name: String, _ container: KeyedDecodingContainer<KeysType>) throws {
-        
-        var attachments = [AttachmentModelType]()
-        
-        for attachmentKey in container.allKeys {
-            
-            let attachmentContainer = try container.nestedContainer(keyedBy: AttachmentModelKeys.self, forKey: attachmentKey)
-            let attachmentTypeString: String? = try attachmentContainer.decodeIfPresent(String.self, forKey: AttachmentModelKeys.type)
-            
-            guard let attachmentType = AttachmentModelTypeKeys(attachmentTypeString) else {
-                
-                throw SkinSlotModelDecodingError.attachmentTypeUnknown
-            }
-            
-            let name = attachmentKey.stringValue
-            
-            switch attachmentType {
-            case .region:
-                attachments.append(AttachmentModelType.region(try RegionAttachmentModel(name, try container.nestedContainer(keyedBy: RegionAttachmentModel.Keys.self, forKey: attachmentKey))))
-                
-            case .boundingBox:
-                attachments.append(AttachmentModelType.boundingBox(try BoundingBoxAttachmentModel(name, try container.nestedContainer(keyedBy: BoundingBoxAttachmentModel.Keys.self, forKey: attachmentKey))))
-                
-            case .mesh:
-                attachments.append(AttachmentModelType.mesh(try MeshAttachmentModel(name, try container.nestedContainer(keyedBy: MeshAttachmentModel.Keys.self, forKey: attachmentKey))))
-                
-            case .linkedMesh:
-                attachments.append(AttachmentModelType.linkedMesh(try LinkedMeshAttachmentModel(name, try container.nestedContainer(keyedBy: LinkedMeshAttachmentModel.Keys.self, forKey: attachmentKey))))
-                
-            case .path:
-                attachments.append(AttachmentModelType.path(try PathAttachmentModel(name, try container.nestedContainer(keyedBy: PathAttachmentModel.Keys.self, forKey: attachmentKey))))
-                
-            case .point:
-                attachments.append(AttachmentModelType.point(try PointAttachmentModel(name, try container.nestedContainer(keyedBy: PointAttachmentModel.Keys.self, forKey: attachmentKey))))
-                
-            case .clipping:
-                attachments.append(AttachmentModelType.clipping(try ClippingAttachmentModel(name, try container.nestedContainer(keyedBy: ClippingAttachmentModel.Keys.self, forKey: attachmentKey))))
-            }
-        }
-        
-        self.name = name
-        self.attachments = attachments
-    }
-}
-
-//MARK: - Attachments
-
-enum AttachmentModelType {
-    
-    case region(RegionAttachmentModel)
-    case boundingBox(BoundingBoxAttachmentModel)
-    case mesh (MeshAttachmentModel)
-    case linkedMesh(LinkedMeshAttachmentModel)
-    case path(PathAttachmentModel)
-    case point(PointAttachmentModel)
-    case clipping(ClippingAttachmentModel)
-
-    var model: AttachmentModel {
-        get {
-            switch self {
-            case .region(let model): return model
-            case .boundingBox(let model): return model
-            case .mesh(let model): return model
-            case .linkedMesh(let model): return model
-            case .path(let model): return model
-            case .point(let model): return model
-            case .clipping(let model): return model
-            }
-        }
-    }
-    
-    var modelName: String {
-        get {
-            switch self {
-            case .region(let model): return model.path ?? model.name
-            case .boundingBox(let model): return model.name
-            case .mesh(let model): return model.path ?? model.name
-            case .linkedMesh(let model): return model.path ?? model.name
-            case .path(let model): return model.name
-            case .point(let model): return model.name
-            case .clipping(let model): return model.name
-            }
-        }
-    }
-}
-
-protocol AttachmentModel {
-
-    var name: String { get }
-    var actualName: String? { get }
-}
-
-enum AttachmentModelKeys: String, CodingKey {
-    
-    case type
-}
-
-enum AttachmentModelTypeKeys: String {
-    
-    case region
-    case boundingBox = "boundingbox"
-    case mesh
-    case linkedMesh = "linkedmesh"
-    case path
-    case point
-    case clipping
-    
-    init?(_ value: String?) {
-        
-        let typeValue = value ?? "region"
-        
-        guard let type = AttachmentModelTypeKeys(rawValue: typeValue) else {
-            
-            return nil
-        }
-        
-        self = type
-    }
-}
-
-//MARK: - Region
-
-struct RegionAttachmentModel: AttachmentModel {
-    
-    let name: String
-    let actualName: String?
-    let path: String?
-    let position: CGPoint
-    let scale: CGVector
-    let rotation: CGFloat
-    let size: CGSize
-    let color: ColorModel
-
-    init(_ name: String, _ actualName: String?, _ path: String?, _ x: CGFloat?, _ y: CGFloat?, _ scaleX: CGFloat?, _ scaleY: CGFloat?, _ rotation: CGFloat?, _ width: CGFloat, _ height: CGFloat, _ color: String?) {
-        
-        self.name = name
-        self.actualName = actualName
-        self.path = path
-        self.position = CGPoint(x: x ?? 0, y: y ?? 0)
-        self.scale = CGVector(dx: scaleX ?? 1.0, dy: scaleY ?? 1.0)
-        self.rotation = rotation ?? 0
-        self.size = CGSize(width: width, height: height)
-        self.color = ColorModel(color ?? "FFFFFFFF")
-    }
-}
-
-extension RegionAttachmentModel: SpineDecodableDictionary {
-
-    enum Keys: String, CodingKey {
-        
-        case name
-        case path
-        case x
-        case y
-        case scaleX
-        case scaleY
-        case rotation
-        case width
-        case height
-        case color
-    }
-    
-    typealias KeysType = Keys
-    
-    init(_ name: String, _ container: KeyedDecodingContainer<KeysType>) throws {
-        
-        let actualName: String? = try container.decodeIfPresent(String.self, forKey: .name)
-        let path: String? = try container.decodeIfPresent(String.self, forKey: .path)
-        let x: CGFloat? = try container.decodeIfPresent(CGFloat.self, forKey: .x)
-        let y: CGFloat? = try container.decodeIfPresent(CGFloat.self, forKey: .y)
-        let scaleX: CGFloat? = try container.decodeIfPresent(CGFloat.self, forKey: .scaleX)
-        let scaleY: CGFloat? = try container.decodeIfPresent(CGFloat.self, forKey: .scaleY)
-        let rotation: CGFloat? = try container.decodeIfPresent(CGFloat.self, forKey: .rotation)
-        let width: CGFloat = try container.decode(CGFloat.self, forKey: .width)
-        let height: CGFloat = try container.decode(CGFloat.self, forKey: .height)
-        let color: String? = try container.decodeIfPresent(String.self, forKey: .color)
-        
-        self.init(name, actualName, path, x, y, scaleX, scaleY, rotation, width, height, color)
-    }
-}
 
 //MARK: - Bounding Box
